bot/pyproject.toml
[project]
name = "front"
version = "0.1.0"
description = ""
authors = [
    {name = "amrzhanseitov",email = "juzloamir@gmail.com"}
]
requires-python = ">=3.14, <3.15"
dependencies = [
    "aiogram (>=3.26.0,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
